model-service/requirements.txt

# Hugging Face Transformers и зависимости
transformers==4.38.2
torch>=1.13
sentencepiece

# PEFT (адаптивные слои LoRA и пр.)
peft==0.8.2  


kafka-python


tqdm